refactor: expenseFixture 패키지 변경
