+++ backend/src/test/java/hanglog/expense/fixture/CityFixture.java
@@ -1,4 +1,4 @@
-package hanglog.trip.fixture;
+package hanglog.expense.fixture;
 
 import hanglog.trip.domain.City;
 import java.math.BigDecimal;
