+++ backend/src/test/java/hanglog/expense/service/ExpenseServiceTest.java
@@ -2,9 +2,9 @@ package hanglog.expense.service;
 
 
 import static hanglog.category.fixture.CategoryFixture.EXPENSE_CATEGORIES;
+import static hanglog.expense.fixture.CityFixture.LONDON;
+import static hanglog.expense.fixture.CityFixture.TOKYO;
 import static hanglog.expense.fixture.CurrencyFixture.DEFAULT_CURRENCY;
-import static hanglog.trip.fixture.CityFixture.LONDON;
-import static hanglog.trip.fixture.CityFixture.TOKYO;
 import static hanglog.trip.fixture.DayLogFixture.EXPENSE_JAPAN_DAYLOG;
 import static hanglog.trip.fixture.DayLogFixture.EXPENSE_LONDON_DAYLOG;
 import static hanglog.trip.fixture.ExpenseFixture.EURO_10000;
