+++ backend/src/test/java/hanglog/trip/service/TripServiceTest.java
@@ -1,8 +1,8 @@
 package hanglog.trip.service;
 
+import static hanglog.expense.fixture.CityFixture.LONDON;
+import static hanglog.expense.fixture.CityFixture.PARIS;
 import static hanglog.global.exception.ExceptionCode.NOT_FOUND_TRIP_ID;
-import static hanglog.trip.fixture.CityFixture.LONDON;
-import static hanglog.trip.fixture.CityFixture.PARIS;
 import static hanglog.trip.fixture.TripFixture.LONDON_TRIP;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
